requirements.txt
fastapi
uvicorn
motor               # ✅ Async MongoDB client (replaces pymongo)
httpx               # ✅ Async HTTP client (replaces requests)

python-dotenv
openai
aiohttp             # ✅ Required for openai>=1.0.0+
fuzzywuzzy
python-Levenshtein>=0.12.0

Jinja2
python-multipart

itsdangerous
pydantic>=2.0
pydantic_ai>=0.1.5
typing_extensions>=4.5
